Module aws_sdk_health::model

Modules

affected_entity

See AffectedEntity

date_time_range

See DateTimeRange

entity_aggregate

See EntityAggregate

entity_filter

See EntityFilter

event

See Event

event_account_filter

See EventAccountFilter

event_aggregate

See EventAggregate

event_description

See EventDescription

event_details

See EventDetails

event_details_error_item

See EventDetailsErrorItem

event_filter

See EventFilter

event_type

See EventType

event_type_filter

See EventTypeFilter

organization_affected_entities_error_item

See OrganizationAffectedEntitiesErrorItem

organization_event

See OrganizationEvent

organization_event_details

See OrganizationEventDetails

organization_event_details_error_item

See OrganizationEventDetailsErrorItem

organization_event_filter

See OrganizationEventFilter

Structs

AffectedEntity

Information about an entity that is affected by a Health event.

DateTimeRange

A range of dates and times that is used by the EventFilter and EntityFilter objects. If from is set and to is set:

EntityAggregate

The number of entities that are affected by one or more events. Returned by the DescribeEntityAggregates operation.

EntityFilter

The values to use to filter results from the EntityFilter

Event

Summary information about an AWS Health event.

EventAccountFilter

The values used to filter results from the DescribeEventDetailsForOrganization and DescribeAffectedEntitiesForOrganization operations.

EventAggregate

The number of events of each issue type. Returned by the DescribeEventAggregates operation.

EventDescription

The detailed description of the event. Included in the information returned by the

EventDetails

Detailed information about an event. A combination of an Event object, an EventDescription object, and additional metadata about the event. Returned by

EventDetailsErrorItem

Error information returned when a DescribeEventDetails operation can't find a specified event.

EventFilter

The values to use to filter results from the DescribeEvents and

EventType

Contains the metadata about a type of event that is reported by AWS Health. The

EventTypeFilter

The values to use to filter results from the DescribeEventTypes

OrganizationAffectedEntitiesErrorItem

Error information returned when a DescribeAffectedEntitiesForOrganization operation can't find or process a

OrganizationEvent

Summary information about an event, returned by the DescribeEventsForOrganization operation.

OrganizationEventDetails

Detailed information about an event. A combination of an Event object, an EventDescription object, and additional metadata about the event. Returned by

OrganizationEventDetailsErrorItem

Error information returned when a DescribeEventDetailsForOrganization operation can't find a specified

OrganizationEventFilter

The values to filter results from the DescribeEventsForOrganization operation.

Enums

EntityStatusCode

NOTE: EntityStatusCode::Unknown has been renamed to ::UnknownValue.

EventAggregateField

EventScopeCode

EventStatusCode

EventTypeCategory