Module aws_sdk_health::operation

Structs

DescribeAffectedAccountsForOrganization

Returns a list of accounts in the organization from AWS Organizations that are affected by the

DescribeAffectedEntities

Returns a list of entities that have been affected by the specified events, based on the

DescribeAffectedEntitiesForOrganization

Returns a list of entities that have been affected by one or more events for one or more

DescribeEntityAggregates

Returns the number of entities that are affected by each of the specified events. If no

DescribeEventAggregates

Returns the number of events of each event type (issue, scheduled change, and account

DescribeEventDetails

Returns detailed information about one or more specified events. Information includes

DescribeEventDetailsForOrganization

Returns detailed information about one or more specified events for one or more AWS

DescribeEventTypes

Returns the event types that meet the specified filter criteria. You can use this API

DescribeEvents

Returns information about events that meet the specified filter criteria. Events are

DescribeEventsForOrganization

Returns information about events across your organization in AWS Organizations. You can use

DescribeHealthServiceStatusForOrganization

This operation provides status information on enabling or disabling AWS Health to work

DisableHealthServiceAccessForOrganization

Disables AWS Health from working with AWS Organizations. To call this operation, you must sign

EnableHealthServiceAccessForOrganization

Enables AWS Health to work with AWS Organizations. You can use the organizational view feature